+++ src/_features.rs
@@ -23,4 +23,4 @@
 //!
 //! * **unstable-replace**: Enable [`Command::replace`](https://github.com/clap-rs/clap/issues/2836)
 //! * **unstable-grouped**: Enable [`ArgMatches::grouped_values_of`](https://github.com/clap-rs/clap/issues/2924)
-//! * **unstable-v4**: Preview features which will be stable on the v4.0 release
+//! * **unstable-v5**: Preview features which will be stable on the v5.0 release