+++ tests/derive/flatten.rs
@@ -20,7 +20,6 @@ use clap::{Args, Parser, Subcommand};
 fn flatten() {
     #[derive(Args, PartialEq, Debug)]
     struct Common {
-        #[clap(value_parser)]
         arg: i32,
     }
 
@@ -45,7 +44,6 @@ fn flatten() {
 fn flatten_twice() {
     #[derive(Args, PartialEq, Debug)]
     struct Common {
-        #[clap(value_parser)]
         arg: i32,
     }
 
@@ -64,13 +62,12 @@ fn flatten_twice() {
 fn flatten_in_subcommand() {
     #[derive(Args, PartialEq, Debug)]
     struct Common {
-        #[clap(value_parser)]
         arg: i32,
     }
 
     #[derive(Args, PartialEq, Debug)]
     struct Add {
-        #[clap(short, action)]
+        #[clap(short)]
         interactive: bool,
         #[clap(flatten)]
         common: Common,
@@ -79,7 +76,7 @@ fn flatten_in_subcommand() {
     #[derive(Parser, PartialEq, Debug)]
     enum Opt {
         Fetch {
-            #[clap(short, action)]
+            #[clap(short)]
             all: bool,
             #[clap(flatten)]
             common: Common,
@@ -108,7 +105,6 @@ fn flatten_in_subcommand() {
 fn update_args_with_flatten() {
     #[derive(Args, PartialEq, Debug)]
     struct Common {
-        #[clap(value_parser)]
         arg: i32,
     }
 
@@ -138,15 +134,13 @@ enum BaseCli {
 
 #[derive(Args, PartialEq, Debug)]
 struct Command1 {
-    #[clap(value_parser)]
     arg1: i32,
-    #[clap(value_parser)]
+
     arg2: i32,
 }
 
 #[derive(Args, PartialEq, Debug)]
 struct Command2 {
-    #[clap(value_parser)]
     arg2: i32,
 }
 
@@ -199,7 +193,6 @@ fn flatten_with_doc_comment() {
     #[derive(Args, PartialEq, Debug)]
     struct Common {
         /// This is an arg. Arg means "argument". Command line argument.
-        #[clap(value_parser)]
         arg: i32,
     }
 
@@ -227,7 +220,7 @@ fn docstrings_ordering_with_multiple_clap() {
     /// This is the docstring for Flattened
     #[derive(Args)]
     struct Flattened {
-        #[clap(long, action)]
+        #[clap(long)]
         foo: bool,
     }
 
@@ -248,7 +241,7 @@ fn docstrings_ordering_with_multiple_clap_partial() {
     /// This is the docstring for Flattened
     #[derive(Args)]
     struct Flattened {
-        #[clap(long, action)]
+        #[clap(long)]
         foo: bool,
     }