+++ tests/derive/raw_bool_literal.rs
@@ -13,9 +13,9 @@ fn raw_bool_literal() {
     #[derive(Parser, Debug, PartialEq)]
     #[clap(name = "raw_bool")]
     struct Opt {
-        #[clap(raw(false), value_parser)]
+        #[clap(raw(false))]
         a: String,
-        #[clap(raw(true), value_parser)]
+        #[clap(raw(true))]
         b: String,
     }